update for annotation indentifinder and update de return's validation method and add method get fails, get fauls

+++ src/main/java/org/doit/project/easyvalidation/annotations/Size.java
@@ -17,9 +17,9 @@ import org.doit.project.easyvalidation.interfaces.impl.SizeValidator;
 public @interface Size {
 	
 	@Property
-	public long	min() default Empty.INTEGER;
+	public int min() default Empty.INTEGER;
 	
 	@Property
-	public long max() default Empty.INTEGER;
+	public int max() default Empty.INTEGER;
 }
